+++ aqua/core/configurer/config_paths.py
@@ -1,0 +1,80 @@
+"""Configuration path helpers for AQUA.
+
+This module resolves the configuration directory/file and the machine name.
+It knows nothing about catalogs - that responsibility lives entirely in
+`config_catalog.ConfigCatalog`, since "catalog" is an intake concept in AQUA.
+"""
+
+import os
+
+from aqua.core.logger import log_configure
+from aqua.core.util.yaml import load_yaml
+
+from .locator import ConfigLocator
+
+
+class ConfigPaths:
+    """
+    Resolves the AQUA configuration directory/file and the machine name.
+    Has no knowledge of catalogs or intake.
+    """
+
+    def __init__(self, configdir=None, filename="config-aqua.yaml", loglevel="warning", locator=None):
+        """
+        Initialize the ConfigPaths instance.
+
+        Args:
+            configdir (str | None): The directory where the configuration file is located.
+                                        If None, it is determined by the `get_config_dir` method.
+            filename (str): The name of the configuration file. Defaults to 'config-aqua.yaml'.
+            loglevel (str): The logging level. Defaults to 'warning'.
+            locator (ConfigLocator | None): An optional ConfigLocator instance.
+        """
+
+        # set up logger
+        self.logger = log_configure(log_level=loglevel, log_name="ConfigPaths")
+
+        # get the configuration directory and its file
+        self.filename = filename
+        if locator is None:
+            locator = ConfigLocator(filename=filename, configdir=configdir, logger=self.logger)
+        self.locator = locator
+        self.configdir = self.locator.configdir
+        self.config_file = self.locator.config_file
+        self.logger.debug("Configuration file found in %s", self.config_file)
+        self.config_dict = load_yaml(self.config_file)
+
+        # get info on machine on init
+        self.machine = self.get_machine()
+
+    def get_config_dir(self):
+        """
+        Return the path to the configuration directory.
+
+        Notes:
+            This method delegates to `ConfigLocator` and is kept for backward
+            compatibility.
+        """
+        return self.locator.configdir
+
+    def get_machine(self):
+        """
+        Extract the name of the machine from the configuration file
+
+        Returns:
+            str | None: resolved machine name from the configuration file, or None when detection fails.
+        """
+        if not os.path.exists(self.config_file):
+            raise FileNotFoundError(f"Cannot find the basic configuration file {self.config_file}!")
+
+        base = load_yaml(self.config_file)
+        # if we do not know the machine we assume is "unknown"
+        machine = "unknown"
+        # if the configuration file has a machine entry, use it
+        if "machine" in base:
+            self.logger.debug("Machine found in configuration file, set to %s", machine)
+            return base["machine"]
+
+        # warning for unknown machine
+        self.logger.warning("No machine entry found in configuration file, set to %s", machine)
+        return machine
