+++ aqua/core/configurer/configpath.py
@@ -2,57 +2,109 @@
 
 `ConfigPath` is kept as a drop-in replacement for the pre-refactor class so
 that existing code (`ConfigPath(...)` used across the codebase) keeps working
-without any changes. Internally it is just `ConfigPaths` (path/config
-resolution) plus a composed `CatalogBrowser` (intake catalog browsing).
+without any changes - including direct attribute access like
+`config_path.catalog`, `config_path.catalog_available`, or
+`config_path.catalog_file`.
 
-New code should prefer using `ConfigPaths` and `CatalogBrowser` directly -
-`ConfigPath` exists purely for migration purposes and could be deprecated
-once call sites have been updated.
+Internally, `ConfigPath` is just `ConfigPaths` (config dir/machine
+resolution) composed with `ConfigCatalog` (all catalog handling, including
+intake). Catalog-related attributes are forwarded via properties so reads
+AND writes (e.g. `config_path.catalog = "some_other_catalog"`) still behave
+exactly as before.
+
+New code should use `ConfigPaths` and `ConfigCatalog` directly. `ConfigPath`
+exists purely for migration and could be deprecated once call sites are
+updated.
 """
 
-from .catalogbrowser import CatalogBrowser
-from .configpaths import ConfigPaths
+from .config_catalog import ConfigCatalog
+from .config_paths import ConfigPaths
 
 
 class ConfigPath(ConfigPaths):
     """
     Backward-compatible class preserving the full original `ConfigPath` API.
-
-    All path/machine-resolution behavior is inherited unchanged from
-    `ConfigPaths`. All intake-catalog behavior (browsing, delivering,
-    scanning, displaying) is delegated to an internal `CatalogBrowser`
-    instance, built against `self` (since `ConfigPath` *is* a `ConfigPaths`).
-
-    Because `CatalogBrowser` mutates `paths.catalog` / `paths.catalog_file` /
-    `paths.machine_file` in place (e.g. in `deliver_intake_catalog`), and
-    `self._browser.paths is self`, those mutations are still visible on the
-    `ConfigPath` instance exactly as before - `config_path.catalog` reflects
-    the last delivered catalog, same as in the original implementation.
     """
 
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self._browser = CatalogBrowser(self)
+    def __init__(self, configdir=None, filename="config-aqua.yaml", catalog=None, loglevel="warning", locator=None):
+        super().__init__(configdir=configdir, filename=filename, loglevel=loglevel, locator=locator)
+        self._catalog = ConfigCatalog(self, catalog=catalog)
+
+    # -- forwarded catalog-related attributes (read + write) --
+
+    @property
+    def catalog(self):
+        return self._catalog.catalog
+
+    @catalog.setter
+    def catalog(self, value):
+        self._catalog.catalog = value
+
+    @property
+    def catalog_available(self):
+        return self._catalog.catalog_available
+
+    @property
+    def base_available(self):
+        return self._catalog.base_available
+
+    @property
+    def catalog_file(self):
+        return self._catalog.catalog_file
+
+    @catalog_file.setter
+    def catalog_file(self, value):
+        self._catalog.catalog_file = value
+
+    @property
+    def machine_file(self):
+        return self._catalog.machine_file
+
+    @machine_file.setter
+    def machine_file(self, value):
+        self._catalog.machine_file = value
+
+    # -- forwarded catalog-related methods --
+
+    def get_catalog(self):
+        """See `ConfigCatalog.get_catalog`."""
+        return self._catalog.get_catalog()
+
+    def get_base(self):
+        """See `ConfigCatalog.get_base`."""
+        return self._catalog.get_base()
+
+    def get_catalog_filenames(self, catalog=None):
+        """See `ConfigCatalog.get_catalog_filenames`."""
+        return self._catalog.get_catalog_filenames(catalog)
+
+    def get_reader_filenames(self, catalog=None):
+        """See `ConfigCatalog.get_reader_filenames`."""
+        return self._catalog.get_reader_filenames(catalog)
+
+    def get_machine_info(self):
+        """See `ConfigCatalog.get_machine_info`."""
+        return self._catalog.get_machine_info()
 
     def browse_catalogs(self, model: str, exp: str, source: str):
-        """See `CatalogBrowser.browse_catalogs`."""
-        return self._browser.browse_catalogs(model, exp, source)
+        """See `ConfigCatalog.browse_catalogs`."""
+        return self._catalog.browse_catalogs(model, exp, source)
 
     def deliver_intake_catalog(self, model, exp, source, catalog=None):
-        """See `CatalogBrowser.deliver_intake_catalog`."""
-        return self._browser.deliver_intake_catalog(model, exp, source, catalog=catalog)
+        """See `ConfigCatalog.deliver_intake_catalog`."""
+        return self._catalog.deliver_intake_catalog(model, exp, source, catalog=catalog)
 
     def scan_catalog(self, cat, model=None, exp=None, source=None):
-        """See `CatalogBrowser.scan_catalog`."""
-        return self._browser.scan_catalog(cat, model=model, exp=exp, source=source)
+        """See `ConfigCatalog.scan_catalog`."""
+        return self._catalog.scan_catalog(cat, model=model, exp=exp, source=source)
 
     def show_catalog_content(self, catalog=None, model=None, exp=None, source=None, verbose=True, show_descriptions=False):
-        """See `CatalogBrowser.show_catalog_content`."""
-        return self._browser.show_catalog_content(
+        """See `ConfigCatalog.show_catalog_content`."""
+        return self._catalog.show_catalog_content(
             catalog=catalog, model=model, exp=exp, source=source, verbose=verbose, show_descriptions=show_descriptions
         )
 
     @staticmethod
     def format_catalog_structure(structure, catalog_name, descriptions=None):
-        """See `CatalogBrowser.format_catalog_structure`."""
-        return CatalogBrowser.format_catalog_structure(structure, catalog_name, descriptions)
+        """See `ConfigCatalog.format_catalog_structure`."""
+        return ConfigCatalog.format_catalog_structure(structure, catalog_name, descriptions)
